backend/pyproject.toml
[project]
name = "anchor-assistant"
version = "0.2.0"
description = "千广传媒主播智能问答助手 - LangChain 增强版"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "chromadb>=0.4.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "python-multipart>=0.0.6",
    "pdfplumber>=0.10.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # LangChain 核心
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "langchain-chroma>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    # 向量存储
    "faiss-cpu>=1.7.4",
    # 混合检索
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.2.0",
    # 文档处理
    "pypdf>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []
